Move beaconcha.in into own directory
