+++ rotkehlchen/chain/ethereum/modules/eth2/eth2.py
@@ -56,7 +56,7 @@ if TYPE_CHECKING:
     from rotkehlchen.db.dbhandler import DBHandler
     from rotkehlchen.db.drivers.gevent import DBCursor
     from rotkehlchen.db.filtering import Eth2DailyStatsFilterQuery
-    from rotkehlchen.externalapis.beaconchain import BeaconChain
+    from rotkehlchen.externalapis.beaconchain.service import BeaconChain
 
 logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
