+++ rotkehlchen/externalapis/beaconchain/constants.py
@@ -1,0 +1,3 @@
+MAX_WAIT_SECS = 60
+BEACONCHAIN_READ_TIMEOUT = 75
+BEACONCHAIN_ROOT_URL = 'https://beaconcha.in'
